amazonka: return Env to its correct home; break module import cycles
